+++ packages/merkleized-metadata/src/MerkleizedMetadata.ts
@@ -16,22 +16,32 @@ import { transformMetadata } from './transform';
 import { ChainInfo, ChainInfoOptional } from './types.js';
 
 /**
- * @name MerkleizedMetatada
+ * Error thrown when decoding extrinsic data fails
+ */
+export class ExtrinsicDecodingError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = 'ExtrinsicDecodingError';
+  }
+}
+
+/**
+ * @name MerkleizedMetadata
  * @description Utility for calculating merkleized metadata hash according to RFC-0078
  */
-export class MerkleizedMetatada {
+export class MerkleizedMetadata {
   readonly #metadata: Metadata;
   readonly #chainInfo: ChainInfo;
 
   /**
-   * Create a new MetatadaMerkleizer instance
+   * Create a new MerkleizedMetadata instance
    *
    * @param metadata - The metadata to calculate hash for
    * @param chainInfo - Chain-specific information
    */
   constructor(metadata: Metadata | HexString | Uint8Array, chainInfo: ChainInfoOptional) {
     // Try decode metadata
-    if (typeof metadata === 'string' || metadata instanceof Uint8Array) {
+    if (!(metadata instanceof Metadata)) {
       metadata = $Metadata.tryDecode(metadata);
     }
 
@@ -70,6 +80,14 @@ export class MerkleizedMetatada {
     return blake3AsU8a($MetadataDigest.encode(digest));
   }
 
+  /**
+   * Decode extrinsic data and collect leaf indices
+   *
+   * @param toDecode - Data to decode
+   * @param typeRefs - Type references to use for decoding
+   * @param typeInfo - Type information
+   * @returns Array of leaf indices
+   */
   #decodeAndCollectLeaves(toDecode: Uint8Array, typeRefs: TypeRef[], typeInfo: TypeInfo[]): number[] {
     type PrimitiveType =
       | 'bool'
@@ -120,6 +138,7 @@ export class MerkleizedMetatada {
       void: $.Null,
     };
 
+    // Create a map of type IDs to their indices in the typeInfo array
     const refIdToIdx = new Map<number, number[]>();
     typeInfo.forEach((one, idx) => {
       const bag = refIdToIdx.get(one.typeId);
@@ -130,21 +149,30 @@ export class MerkleizedMetatada {
       }
     });
 
+    // Helper function to decode data
     const decode = ($codec: $.AnyShape) => {
-      const decoded = $codec.decode(toDecode) as any;
-
-      // @ts-ignore
-      const encodedLength = $codec.encode(decoded).length;
-      toDecode = toDecode.subarray(encodedLength);
-
-      return decoded;
+      try {
+        const decoded = $codec.decode(toDecode) as any;
+        // @ts-ignore
+        const encodedLength = $codec.encode(decoded).length;
+        toDecode = toDecode.subarray(encodedLength);
+        return decoded;
+      } catch (error: any) {
+        throw new ExtrinsicDecodingError(`Failed to decode data: ${error.message || String(error)}`);
+      }
     };
 
     const collectedIndices = new Set<number>();
 
+    // Recursive function to decode and collect leaf indices
     const decodeAndCollect = (one: TypeRef) => {
       if (one.type === 'perId') {
-        const indexes = refIdToIdx.get(one.value)!;
+        const indexes = refIdToIdx.get(one.value);
+
+        if (!indexes || indexes.length === 0) {
+          throw new ExtrinsicDecodingError(`Type ID ${one.value} not found in type info`);
+        }
+
         const [idx] = indexes;
 
         if (indexes.length === 1) collectedIndices.add(idx);
@@ -157,41 +185,60 @@ export class MerkleizedMetatada {
             for (let i = 0; i < length; i += 1) {
               decodeAndCollect(typeDef.value);
             }
-
             break;
+
           case 'tuple':
             typeDef.value.forEach(decodeAndCollect);
             break;
+
           case 'array':
             for (let i = 0; i < typeDef.value.len; i += 1) {
               decodeAndCollect(typeDef.value.typeParam);
             }
             break;
+
           case 'composite':
             typeDef.value.forEach((one) => decodeAndCollect(one.ty));
             break;
+
           case 'bitSequence':
+            // BitSequence doesn't need further decoding
             break;
+
           case 'enumeration':
             const selectedIdx = decode($.u8);
-            const [{ fields }, idx] = refIdToIdx
-              .get(one.value)!
-              .map((id) => [typeInfo[id].typeDef.value, id] as [EnumerationVariant, number])!
-              .find(([{ index }]) => index === selectedIdx)!;
 
+            const variantInfo = refIdToIdx
+              .get(one.value)
+              ?.map((id) => [typeInfo[id].typeDef.value, id] as [EnumerationVariant, number])
+              .find(([{ index }]) => index === selectedIdx);
+
+            if (!variantInfo) {
+              throw new ExtrinsicDecodingError(
+                `Enum variant with index ${selectedIdx} not found for type ID ${one.value}`,
+              );
+            }
+
+            const [{ fields }, idx] = variantInfo;
             collectedIndices.add(idx);
             fields.forEach(({ ty }) => decodeAndCollect(ty));
             break;
+
+          default:
+            // This should never happen as we've covered all possible typeDef types
+            throw new ExtrinsicDecodingError(`Unsupported type definition: ${(typeDef as any).type}`);
         }
       } else {
         decode(primitiveCodecs[one.type]);
       }
     };
 
-    typeRefs.map(decodeAndCollect);
+    // Process all type references
+    typeRefs.forEach(decodeAndCollect);
 
+    // Check if there are any remaining bytes
     if (toDecode.length > 0) {
-      throw new Error('Extra bytes at the end of the extrinsic!');
+      throw new ExtrinsicDecodingError(`Extra bytes at the end of the extrinsic: ${toDecode.length} bytes remaining`);
     }
 
     return [...collectedIndices].sort((a, b) => a - b);
@@ -205,8 +252,7 @@ export class MerkleizedMetatada {
    * @returns The metadata proof
    */
   proofForExtrinsic(extrinsic: Uint8Array | HexString, additionalSigned?: Uint8Array | HexString): Uint8Array {
-    // In a real implementation, we would:
-    // 1. Decode the extrinsic to extract call data, extrinsic extra, and signed extra
+    // Decode the extrinsic to extract call data, extrinsic extra, and signed extra
     const $Codec = $.Tuple($.compactU32, $ExtrinsicVersion, $.RawHex);
 
     const [, version, bytes] = $Codec.tryDecode(extrinsic);
@@ -220,7 +266,7 @@ export class MerkleizedMetatada {
       `Invalid extrinsic version, expected version ${extrinsicMetadata.version}`,
     );
 
-    // 2. Identify the type IDs used in the extrinsic
+    // Identify the type IDs used in the extrinsic
     const typeRefs: TypeRef[] = [];
     if (version.signed) {
       typeRefs.push(
@@ -253,21 +299,28 @@ export class MerkleizedMetatada {
 
   /**
    * Generate proof for extrinsic parts
+   *
+   * @param callData - Call data
+   * @param includedInExtrinsic - Data included in extrinsic
+   * @param includedInSignedData - Data included in signed data
+   * @returns The metadata proof
    */
   proofForExtrinsicParts(
     callData: Uint8Array | HexString,
     includedInExtrinsic: Uint8Array | HexString,
     includedInSignedData: Uint8Array | HexString,
   ): Uint8Array {
-    const payload = concatU8a(
-      toU8a(callData), // prettier-end-here
-      toU8a(includedInExtrinsic),
-      toU8a(includedInSignedData),
-    );
+    const payload = concatU8a(toU8a(callData), toU8a(includedInExtrinsic), toU8a(includedInSignedData));
 
     return this.proofForExtrinsicPayload(payload);
   }
 
+  /**
+   * Generate proof for extrinsic payload
+   *
+   * @param txPayload - Transaction payload
+   * @returns The metadata proof
+   */
   proofForExtrinsicPayload(txPayload: Uint8Array | HexString): Uint8Array {
     const { extrinsicMetadata, typeInfo } = transformMetadata(this.#metadata);
 
@@ -288,6 +341,13 @@ export class MerkleizedMetatada {
     });
   }
 
+  /**
+   * Look up a constant in the metadata
+   *
+   * @param pallet - Pallet name
+   * @param constant - Constant name
+   * @returns Constant value
+   */
   #lookupConstant<T extends any = any>(pallet: string, constant: string): T {
     const registry = new PortableRegistry(this.#metadata.latest);
     const targetPallet = this.#metadata.latest.pallets.find((p) => stringCamelCase(p.name) === pallet);
