+++ packages/merkleized-metadata/src/merkle/index.ts
@@ -1,0 +1,2 @@
+export { buildMerkleTree } from './tree';
+export { generateProof, verifyProof } from './proof';
