+++ packages/merkleized-metadata/src/transform/index.ts
@@ -1,0 +1,11 @@
+export { getAccessibleTypes } from './accessibleTypes';
+export { transformMetadata } from './metadata';
+export {
+  PRIMITIVE_TYPE_MAP,
+  COMPACT_TYPE_MAP,
+  isPrimitiveType,
+  getPrimitiveTypeTag,
+  isCompactType,
+  getCompactTypeTag,
+  isVoidType,
+} from './typeUtils';
